# server/docs/swagger.yaml
basePath: /ttms
definitions:
  common.State:
    properties:
      data:
        description: 失败时返回空
      status_code:
        description: 状态码，0-成功，其他值-失败
        type: integer
      status_msg:
        description: 返回状态描述
        type: string
    type: object
...
  db.GetMoviesByIDsRow:
    properties:
      actors:
        items:
          type: string
        type: array
      alias_name:
        type: string
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      score:
        type: number
    type: object
  db.GetMoviesByNameOrContentRow:
    properties:
      actors:
        items:
          type: string
        type: array
      alias_name:
        type: string
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      score:
        type: number
      total:
        type: integer
    type: object
  db.GetMoviesOrderByBoxOfficeRow:
    properties:
      actors:
        items:
          type: string
        type: array
      alias_name:
        type: string
      avatar:
        type: string
      box_office:
        type: number
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      score:
        type: number
    type: object
  db.GetMoviesOrderByUserMovieCountRow:
    properties:
      actors:
        items:
          type: string
        type: array
      alias_name:
        type: string
      avatar:
        type: string
      fallow_count:
        type: integer
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      score:
        type: number
    type: object
  db.GetMoviesOrderByVisitCountRow:
    properties:
      actors:
        items:
          type: string
        type: array
      alias_name:
        type: string
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      score:
        type: number
      total:
        type: integer
    type: object
...
  reply.GetAreas:
    properties:
      list:
        items:
          type: string
        type: array
    type: object
...
  reply.GetMovieByID:
    properties:
      actors:
        items:
          type: string
        type: array
      alias_name:
        type: string
      area:
        type: string
      avatar:
        type: string
      box_office:
        type: number
      comment_count:
        type: integer
      content:
        type: string
      director:
        type: string
      duration:
        type: integer
      follow_count:
        type: integer
      id:
        type: integer
      isComment:
        type: boolean
      isFollow:
        type: boolean
      name:
        type: string
      period:
        type: string
      score:
        type: number
      tags:
        items:
          type: string
        type: array
      visit_count:
        type: integer
    type: object
  reply.GetMovies:
    properties:
      list:
        items:
          $ref: '#/definitions/reply.GetMoviesWithTags'
        type: array
    type: object
  reply.GetMoviesByNameOrContent:
    properties:
      list:
        items:
          $ref: '#/definitions/db.GetMoviesByNameOrContentRow'
        type: array
    type: object
  reply.GetMoviesByTagPeriodArea:
    properties:
      list:
        items:
          $ref: '#/definitions/reply.GetMoviesByTagPeriodAreaRow'
        type: array
    type: object
  reply.GetMoviesByTagPeriodAreaRow:
    properties:
      actors:
        items:
          type: string
        type: array
      alias_name:
        type: string
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      score:
        type: number
      total:
        type: integer
    type: object
  reply.GetMoviesOrderByBoxOffice:
    properties:
      list:
        items:
          $ref: '#/definitions/db.GetMoviesOrderByBoxOfficeRow'
        type: array
    type: object
  reply.GetMoviesOrderByRecentVisitNum:
    properties:
      list:
        items:
          $ref: '#/definitions/db.GetMoviesByIDsRow'
        type: array
    type: object
  reply.GetMoviesOrderByUserMovieCount:
    properties:
      list:
        items:
          $ref: '#/definitions/db.GetMoviesOrderByUserMovieCountRow'
        type: array
    type: object
  reply.GetMoviesOrderByVisitCount:
    properties:
      list:
        items:
          $ref: '#/definitions/db.GetMoviesOrderByVisitCountRow'
        type: array
    type: object
  reply.GetMoviesWithTags:
    properties:
      actors:
        items:
          type: string
        type: array
      alias_name:
        type: string
      area:
        type: string
      avatar:
        type: string
      content:
        type: string
      director:
        type: string
      duration:
        type: integer
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      tags:
        items:
          type: string
        type: array
      total:
        type: integer
    type: object
...
  request.CreateMovie:
    properties:
      actors:
        description: 演员列表
        items:
          maxLength: 5
          type: string
        minItems: 1
        type: array
      alias_name:
        description: 别名
        minLength: 1
        type: string
      area:
        maxLength: 20
        minLength: 1
        type: string
      avatar:
        description: 图像
        type: string
      content:
        description: 简介
        maxLength: 1000
        minLength: 1
        type: string
      director:
        description: 导演
        minLength: 1
        type: string
      duration:
        description: 时长(分钟)
        minimum: 1
        type: integer
      name:
        minLength: 1
        type: string
      period:
        description: 上映时间 秒时间戳
        minimum: 0
        type: integer
      tags:
        description: 标签，不能有重复的
        items:
          maxLength: 5
          type: string
        minItems: 1
        type: array
    required:
    - actors
    - alias_name
    - area
    - content
    - director
    - duration
    - name
    - period
    - tags
    type: object
...
  request.DeleteMovieByID:
    properties:
      movie_id:
        minimum: 1
        type: integer
    required:
    - movie_id
    type: object
...
  request.UpdateMovie:
    properties:
      actors:
        description: 演员列表
        items:
          maxLength: 5
          type: string
        minItems: 1
        type: array
      alias_name:
        description: 别名
        minLength: 1
        type: string
      area:
        type: string
      avatar:
        type: string
      content:
        minLength: 1
        type: string
      director:
        minLength: 1
        type: string
      id:
        description: 电影ID
        type: integer
      name:
        minLength: 1
        type: string
      period:
        description: 上映时间
        type: integer
    required:
    - actors
    - alias_name
    - area
    - content
    - director
    - id
    - name
    - period
    type: object
...
host: ttms.humraja.xyz
info:
  contact: {}
  description: ttms影院管理系统
  license:
    name: raja,moonman
  title: ttms
  version: "1.0"
paths:
...
  /movie/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 电影信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateMovie'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.State'
      security:
      - BasicAuth: []
      summary: 创建电影
      tags:
      - movie
  /movie/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 电影ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteMovieByID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.State'
      security:
      - BasicAuth: []
      summary: 删除电影(不能删除存在演出计划的电影)
      tags:
      - movie
  /movie/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - in: query
        minimum: 1
        name: movie_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetMovieByID'
              type: object
      security:
      - BasicAuth: []
      summary: 获取电影详细信息
      tags:
      - movie
  /movie/list/areas:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 第几页
        in: query
        name: page
        type: integer
      - default: 5
        description: 每页大小
        in: query
        maximum: 10
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetAreas'
              type: object
      security:
      - BasicAuth: []
      summary: 获得所有存在的地区
      tags:
      - movie
  /movie/list/box_office:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        maximum: 5
        minimum: 1
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetMoviesOrderByBoxOffice'
              type: object
      security:
      - BasicAuth: []
      summary: 根据票房获取固定数量的电影列表(非实时)
      tags:
      - movie
  /movie/list/info:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 第几页
        in: query
        name: page
        type: integer
      - default: 5
        description: 每页大小
        in: query
        maximum: 10
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetMovies'
              type: object
      security:
      - BasicAuth: []
      summary: 分页获取电影的创建时信息
      tags:
      - movie
  /movie/list/key:
    get:
      consumes:
      - application/json
      parameters:
      - description: 关键字
        in: query
        maxLength: 20
        minLength: 1
        name: key
        required: true
        type: string
      - default: 1
        description: 第几页
        in: query
        name: page
        type: integer
      - default: 5
        description: 每页大小
        in: query
        maximum: 10
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetMoviesByNameOrContent'
              type: object
      security:
      - BasicAuth: []
      summary: 通过关键字对电影名,别名以及分页进行模糊查询返回匹配电影的简要信息
      tags:
      - movie
  /movie/list/recent_visit_count:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 第几页
        in: query
        name: page
        type: integer
      - default: 5
        description: 每页大小
        in: query
        maximum: 10
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetMoviesOrderByRecentVisitNum'
              type: object
      security:
      - BasicAuth: []
      summary: 根据近期访问量获取电影列表
      tags:
      - movie
  /movie/list/tag_period_area:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: area
        type: string
      - in: query
        name: end_time
        type: integer
      - enum:
        - period
        - visit_count
        - score
        in: query
        name: order_by
        required: true
        type: string
      - default: 1
        description: 第几页
        in: query
        name: page
        type: integer
      - default: 5
        description: 每页大小
        in: query
        maximum: 10
        name: page_size
        type: integer
      - in: query
        name: start_time
        type: integer
      - in: query
        name: tag_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetMoviesByTagPeriodArea'
              type: object
      security:
      - BasicAuth: []
      summary: 分页查询经过对标签上映时间地区进行筛选并按照指定数据排序的电影
      tags:
      - movie
  /movie/list/user_movie_count:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        maximum: 5
        minimum: 1
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetMoviesOrderByUserMovieCount'
              type: object
      security:
      - BasicAuth: []
      summary: 根据电影期待数获取固定数量的电影列表(非实时)
      tags:
      - movie
  /movie/list/visit_count:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 第几页
        in: query
        name: page
        type: integer
      - default: 5
        description: 每页大小
        in: query
        maximum: 10
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetMoviesOrderByVisitCount'
              type: object
      security:
      - BasicAuth: []
      summary: 根据访问量获取电影列表
      tags:
      - movie
  /movie/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新的电影信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateMovie'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.State'
      security:
      - BasicAuth: []
      summary: 更新电影信息(时长不可更改)
      tags:
      - movie
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"